Proxies (now unneeded), split AlembicConfig.
Update the /research command to also show achieved advancements.

Update the reload command for ResearchHandler to take the server as a
parameter. For the most part this allows us to reload the listeners so
that events continue to fire properly for generated advancements.

For the resource reload listener, presume we're running on an integrated
server and that accessing the server via the DimensionManager would be
sufficient.

It's probably not.